Alteracao nos controllers, melhora nas responses das rotas, criacao das rotas de agendamento e validaçao de nivel de acesso

+++ backend/src/controllers/servicoController.js
@@ -7,25 +7,58 @@ router.use(authMiddleware);
 const Servico = require("../models/Servico");
 
 router.get("/", async (req, res) => {
-    var servicos = await Servico.findAll();
+    try {
 
-    if (servicos.length === 0){
-        return res.status(400).send("Serviços não cadastrados");
+        var servicos = await Servico.findAll();
+
+        if (servicos.length === 0){
+            return res.status(400).send("Serviços não cadastrados");
+        }
+
+        return res.status(200).json(servicos.body);
+    
+    } catch(error){
+        console.log(error);
+        res.status(400).send("Erro na requisição");
     }
-
-    return res.status(200).json(servicos.body);
 });
 
 router.post("/", async (req, res) => {
     
-    await Servico.create(req.body);
-    return res.status(200).send("Serviço cadastrado com sucesso");
+    try {
+
+        var user = await User.findByPk(req.userId);
+
+        if (!user.nivel_acesso === constantes.NIVEL_ACESSO.PROPRIETARIO){
+            return res.status(400).send("Nível de acesso insuficiente");
+        }
+
+        await Servico.create(req.body);
+        return res.status(200).send("Serviço cadastrado com sucesso");
+
+    } catch(error){
+        console.log(error);
+        res.status(400).send("Erro na requisição");
+    }
 });
 
 router.delete("/:id", async (req, res) => {
     
-    await Servico.destroy({where: {id}});
-    return res.status(200).send("Serviço removido com sucesso");
+    try {
+
+        var user = await User.findByPk(req.userId);
+
+        if (!user.nivel_acesso === constantes.NIVEL_ACESSO.PROPRIETARIO){
+            return res.status(400).send("Nível de acesso insuficiente");
+        }
+
+        await Servico.destroy({where: {id}});
+        return res.status(200).send("Serviço removido com sucesso");
+    
+    } catch(error){
+        console.log(error);
+        res.status(400).send("Erro na requisição");
+    }
 });
 
 module.exports = app => app.use("/servicos", router);
